add local object store commands
