+++ lean/commands/object_store/delete.py
@@ -11,19 +11,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from click import command, argument
-
+from click import command
 from lean.click import LeanCommand
 from lean.container import container
+from lean.components.util.object_store_helper import open_storage_directory_in_explorer
 
 
 @command(cls=LeanCommand)
-@argument("key", type=str)
-def delete(key: str) -> str:
+def delete() -> str:
     """
-    Delete a value from the organization's object store.
+    Opens the local storage directory in the file explorer.
     
     """
-    organization_id = container.organization_manager.try_get_working_organization_id()
-    api_client = container.api_client
-    api_client.object_store.delete(key, organization_id)
+    open_storage_directory_in_explorer(container.lean_config_manager)
